+++ src/permutation.c
@@ -8,7 +8,7 @@
 // The initial permutation - IP.
 // 
 // As of now, it is hard-coded.
-static uint8_t IP[] = 
+static int IP[] = 
 {
     58, 50, 42, 34, 26, 18, 10, 2,
     60, 52, 44, 36, 28, 20, 12, 4,
@@ -25,7 +25,7 @@ Perm *get_default_initial_permutation() {
     Perm *p = malloc(sizeof(Perm));
     p->size = 64;
 
-    p->data = malloc(sizeof(uint8_t) * 64);
+    p->data = malloc(sizeof(int) * 64);
 
     for (int i = 0; i < 64; i++) {
         p->data[i] = IP[i];
@@ -37,7 +37,7 @@ Perm *get_default_initial_permutation() {
 Perm *create_permutation(int size, ...) {
     Perm *p = malloc(sizeof(Perm));
     p->size = size;
-    p->data = calloc(size, sizeof(uint8_t));
+    p->data = calloc(size, sizeof(int));
     
     va_list va;
     va_start(va, size);
@@ -64,7 +64,7 @@ void free_permutation(Perm **p) {
 Perm *compute_inverse_permutation(Perm *perm) {
     Perm *p = malloc(sizeof(Perm));
     p->size = perm->size;
-    p->data = calloc(perm->size, sizeof(uint8_t));
+    p->data = calloc(perm->size, sizeof(int));
     
     for (int i = 0; i < perm->size; i++) {
         p->data[perm->data[i]-1] = i+1;
